requirements.txt
tensorflow==2.13.0
opencv-python==4.8.1.78
scikit-learn==1.3.0
streamlit==1.25.0
pillow==9.5.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
